feat: 시간표 학점 반환시 채플은 0.5학점으로 계산하여 반환. 자료형 double로 변경

+++ src/main/kotlin/com/yourssu/soongpt/domain/timetable/business/dto/TimetableResponse.kt
@@ -6,7 +6,7 @@ data class TimetableResponse(
     val timetableId: Long,
     val tag: String,
     var score: Int?,
-    val totalCredit: Int,
+    val totalCredit: Double,
     val courses: List<TimetableCourseResponse>,
 ) {
     companion object {
